requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.11
langchain-community>=0.0.13
pydantic>=2.5.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
pypdf2>=3.0.1
pdfplumber>=0.10.3
streamlit>=1.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.26.3
pytest>=7.4.4
pytest-cov>=4.1.0